refactor: 과도한 도메인 분리 방지를 위해 로또 번호들과 로또 게임을 병합.
