+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -11,6 +11,7 @@ import java.util.Set;
 import java.util.stream.Stream;
 
 import static lotto.domain.LottoNumber.from;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("LottoGame 테스트")
@@ -19,11 +20,9 @@ class LottoGameTest {
     @MethodSource("provideSource_new_정상")
     @DisplayName("new_정상")
     void new_정상(Set<LottoNumber> values) {
-        // Given
-        LottoNumbers lottoNumbers = new LottoNumbers(values);
 
         // When, Then
-        assertDoesNotThrow(() -> new LottoGame(lottoNumbers));
+        assertDoesNotThrow(() -> new LottoGame(values));
     }
 
     static Stream<Arguments> provideSource_new_정상() {
@@ -31,6 +30,26 @@ class LottoGameTest {
                 Arguments.of(new HashSet(Arrays.asList(from(1), from(2), from(3), from(4), from(5), from(6)))),
                 Arguments.of(new HashSet(Arrays.asList(from(11), from(12), from(13), from(14), from(15), from(16)))),
                 Arguments.of(new HashSet(Arrays.asList(from(21), from(22), from(23), from(24), from(25), from(26))))
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("provideSource_new_예외")
+    @DisplayName("new_예외")
+    void new_예외(Set<LottoNumber> values) {
+        // When, Then
+        assertThatExceptionOfType(IllegalArgumentException.class)
+                .isThrownBy(() -> new LottoGame(values));
+    }
+
+    static Stream<Arguments> provideSource_new_예외() {
+        return Stream.of(
+                Arguments.of(new HashSet(Arrays.asList(from(1)))),
+                Arguments.of(new HashSet(Arrays.asList(from(1), from(2)))),
+                Arguments.of(new HashSet(Arrays.asList(from(1), from(2), from(3)))),
+                Arguments.of(new HashSet(Arrays.asList(from(1), from(2), from(3), from(4)))),
+                Arguments.of(new HashSet(Arrays.asList(from(1), from(2), from(3), from(4), from(5)))),
+                Arguments.of(new HashSet(Arrays.asList(from(1), from(2), from(3), from(4), from(5), from(6), from(7))))
         );
     }
